+++ noxfile.py
@@ -117,6 +117,7 @@ class ReverseDependency:
             install_target=".[dev]",
         ),
         ReverseDependency("https://github.com/TeamGraphix/graphix-ibmq"),
+        ReverseDependency("https://github.com/qat-inria/graphix-stim-compiler"),
     ],
 )
 def tests_reverse_dependencies(session: Session, package: ReverseDependency) -> None: